Pulled recipe-list tag cloud into separate component
